#1333 Publish from collaborator modal (#1343)
