+++ ui/src/js/redux/actions/shared/collaborators/collaborators.js
@@ -14,3 +14,8 @@ export const setCanManageCollaborators = canManageCollaborators => dispatcher(
   types.SET_CANMANAGECOLLABORATORS,
   { canManageCollaborators },
 );
+
+export const setPublishFromCollaborators = publishFromCollaborators => dispatcher(
+  types.SET_PUBLISH_FROM_COLLABORATORS,
+  { publishFromCollaborators },
+);
