+++ ui/src/js/redux/constants/constants.js
@@ -62,6 +62,7 @@ export const UPDATE_TRANSITION_STATE = 'UPDATE_TRANSITION_STATE';
  */
 export const SET_COLLABORATORS = 'SET_COLLABORATORS';
 export const SET_CANMANAGECOLLABORATORS = 'SET_CANMANAGECOLLABORATORS';
+export const SET_PUBLISH_FROM_COLLABORATORS = 'SET_PUBLISH_FROM_COLLABORATORS';
 
 
 /**
