+++ ui/src/js/redux/reducers/shared/collaborators/collaborators.js
@@ -1,10 +1,10 @@
 import * as types from 'JS/redux/constants/constants';
 
-
 export default (
   state = {
     collaborators: null,
     canManageCollaborators: null,
+    publishFromCollaborators: false,
   },
   action,
 ) => {
@@ -13,10 +13,17 @@ export default (
       ...state,
       collaborators: action.payload.collaborators,
     };
-  } if (action.type === types.SET_CANMANAGECOLLABORATORS) {
+  }
+  if (action.type === types.SET_CANMANAGECOLLABORATORS) {
     return {
       ...state,
       canManageCollaborators: action.payload.canManageCollaborators,
+    };
+  }
+  if (action.type === types.SET_PUBLISH_FROM_COLLABORATORS) {
+    return {
+      ...state,
+      publishFromCollaborators: action.payload.publishFromCollaborators,
     };
   }
   return state;
